implemented bullet-enemy collision detection and handling (TODO: implement knockback. forgot about that but im sleepy)

+++ core/src/sophomoreproject/game/gameobjects/enemystuff/EnemyInfo.java
@@ -13,7 +13,7 @@ public class EnemyInfo {
     public float size = 1;
     public float attackDamage = 1;
     public float attackDelay = 1.5f;
-    public float health = 5;
+    public int health = 5;
 
     public EnemyInfo() {}
